feat: ajout de title manquants
